Carve out association_reflection_extension
